Updated StressAppTest wiki article (#596)
* Update stressapptest.md

* Modified terminal look

* Changed stressapptest from hidden to not hidden

Also moved it out of memtest folder to live sessions... Why was it part of memtest?

* Moved it back cuz azure no likey me

Yes IK all I need to change was the link in the memtest area, but now that I think about it, it does seem to maybe fit memtest more than anything else so ehh... I guess I will leave it back where it was.

* Slight reformatting